+++ app/src/main/java/org/simple/clinic/selectcountry/UiActions.kt
@@ -2,5 +2,5 @@ package org.simple.clinic.selectcountry
 
 interface UiActions {
   fun goToStateSelectionScreen()
-  fun replaceCurrentScreenToStateSelectionScreen()
+  fun replaceWithStateSelectionScreen()
 }