+++ app/src/test/java/org/simple/clinic/selectcountry/SelectCountryUpdateTest.kt
@@ -147,7 +147,7 @@ class SelectCountryUpdateTest {
         .whenEvent(CountrySaved)
         .then(assertThatNext(
             hasNoModel(),
-            hasEffects(ReplaceCurrentScreenWithStateSelectionScreen)
+            hasEffects(ReplaceWithStateSelectionScreen)
         ))
   }
 }